refactor(sidecar): cleanup between different bls signers
