+++ bolt-sidecar/src/signer/mod.rs
@@ -1,11 +1,11 @@
-use keystore::KeystoreSigner;
-use local::LocalSigner;
-
-use crate::CommitBoostSigner;
-
 pub mod commit_boost;
+use commit_boost::CommitBoostSigner;
+
 pub mod keystore;
+use keystore::KeystoreSigner;
+
 pub mod local;
+use local::LocalSigner;
 
 #[derive(Debug, Clone)]
 /// Signer for BLS signatures.
